+++ src/setDay/test.js
@@ -49,6 +49,20 @@ describe('setDay', function() {
     assert.deepEqual(result, new Date(2014, 8 /* Sep */, 7))
   })
 
+  it('specifies Monday as the first day of the week', function() {
+    var result = setDay(new Date(2014, 8 /* Sep */, 6), 1, {
+      weekStartsOn: 1
+    })
+    assert.deepEqual(result, new Date(2014, 8 /* Sep */, 1))
+  })
+
+  it('specifies Tuesday as the first day of the week', function() {
+    var result = setDay(new Date(2014, 8 /* Sep */, 6), 1, {
+      weekStartsOn: 2
+    })
+    assert.deepEqual(result, new Date(2014, 8 /* Sep */, 8))
+  })
+
   context('the day index is more than 6', function() {
     it('sets the day of the next week', function() {
       var result = setDay(new Date(2014, 8 /* Sep */, 1), 8)
@@ -59,14 +73,14 @@ describe('setDay', function() {
       var result = setDay(new Date(2014, 8 /* Sep */, 1), 7, {
         weekStartsOn: 1
       })
-      assert.deepEqual(result, new Date(2014, 8 /* Sep */, 14))
+      assert.deepEqual(result, new Date(2014, 8 /* Sep */, 8))
     })
 
     it('sets the day of another week in the future', function() {
       var result = setDay(new Date(2014, 8 /* Sep */, 1), 14, {
         weekStartsOn: 1
       })
-      assert.deepEqual(result, new Date(2014, 8 /* Sep */, 21))
+      assert.deepEqual(result, new Date(2014, 8 /* Sep */, 15))
     })
   })
 
@@ -80,14 +94,14 @@ describe('setDay', function() {
       var result = setDay(new Date(2014, 8 /* Sep */, 1), -7, {
         weekStartsOn: 1
       })
-      assert.deepEqual(result, new Date(2014, 7 /* Aug */, 31))
+      assert.deepEqual(result, new Date(2014, 7 /* Aug */, 25))
     })
 
     it('set the day of another week in the past', function() {
       var result = setDay(new Date(2014, 8 /* Sep */, 1), -14, {
         weekStartsOn: 1
       })
-      assert.deepEqual(result, new Date(2014, 7 /* Aug */, 24))
+      assert.deepEqual(result, new Date(2014, 7 /* Aug */, 18))
     })
   })